Initial label creation for a given albumId
Take an album ID from deezer, pull down the album information and then
render the album art along with the name of the artist, and title of the
album. This thing will definitely include some more grafics to make it
look more attractive to kids, but this is a nice start.