Introduce buntdb as the data backend
Store all track information in a buntdb instance, for speed and ease of
use. This means that I only have to get the ID parsing of the cards
working, and I don't need to try to store the reduced data structures on
the low memory cards. It also makes managing the data (when needed) lots
easier, because no actual re-coding of cards will be needed.