Make labels more attractive
Switch the artist and album title. The title would be more important,
and the artist, so putting that first and then including the artist as
secondary information. Adding a number of frames that were fetched from
https://www.vecteezy.com/vector-art/198818-funcky-frames-borders-2-vectors
and then modified to be perfectly square to fit the album art. Now the
labels will have one of the seven custom frames, and the text has a dark
outline with a random color. The colors should probably be expanded a
little to include a few more ones, but overall, this is pretty nice as
it is.