CLI, reworks, etc
Lots of work on the CLI and provisioning of data, as well as the even
streams for seen NFC cards during the TS hackathon. There is a lot of
cleanup to be done, but this version seems to have most (all?)
individual parts working now to at least some degree.

The even streams should be possible to simplify a lot and when combined
with the same type of in streams from the buttons, it should be possible
to simplify even more. Needs a bit of designing of the events so that
they make more sense, and after that it should be fairly straight
forward.

Everything has been mounted inside the tiger box as well. Stuff is
progressing :)