Connect card to sonos player
Now finally got the card reader integrated with the sonos interface, so
that when a card is added, the songs start playing. Made the debug
logging possible to turn on by issuing a --debug flag.

The card bouncing problem seems to have been solved by lowering the
max speed of the SPI bus, but that remains to be seen. Multiple tests
performed without any real read errors though.