Add card reading to commands
Add reading of card ids for dumping card info, adding albums or
rendering labels. Refactor some of the card reader related code to put
it all in the nfc module and remove some dead code from there pertaining
to the IRQ pin that is currently not in use (just couldn't get it
properly working). Drop the playlist command from the CLI for now.