Render sheets of 9 labels
Add a switch for rendering all labels in the database including cut
marks. This allows all of them to be printed in one go at 600 dpi,
without actually rendering them individually. Also add a single command
for dumping the album info about a card to easily identify which one it
it.