Add support for deezer playlists
Support playlists as well as albums from deezer. Introduces an
abstraction `TrackList` that can be used for both the albums and the
playlists to get tracks. Since the sonos package operates on the tracks
only, this doesn't need to be touched. Label generation, search, and
card adding have all been updated to support playlists as well.