Add album/playlist refresh option
It seems like the track id's on deezer changes from time to time even
for albums. Also playlists might change, and would need periodic
refreshing to show the reflected changes. Add a --refresh flag to the
start command that refreshes the content of the card database on
startup, and then every 24h afterwards. This way the card database will
always be fairly up to date.