Make refresh only affect playlists
It actually seems like the album problems stem from the entire album ID
changing, so it doesn't help to update the album entries. For playlists
this is still valid, so keeping that functionality.

Add check command to check all the entries, so see of they are pointing
to existing IDs. Check could also be extended to check that all the
tracks for a give album/playlist are still valid and/or the same.