Make server runnable locally
Add mock packages for the UI and NFC that are otherwize hardware
dependent on running on the tiger player. Now, it's also possible to
build the thing locally by loading the mock packages instead of the real
ones.

Also add a prometheus server so that it will be possible to produce some
juicy metrics for the player afterwards.