Queue albums/playlists instead of tracks
Queueing individual tracks is very slow. On larger playlists this can
take 30s, whereas just adding a single URI that represents the playlist
will take less than a second. Stole some ideas around what the request
should look like from socos for this, and it seems to work for the tiger
player as well.

Using the album/playlist ID instead of the tracks, also makes it
redundant save the tracks in the database, but this will be kept in just
since we get it for free during the provisioning anyway.