Set queue when playing
The "source" needs to be set up properly before starting to play
something on the sonos speakers. This was missed because once set up, I
don't change it, so it only applies in our house when the speaker has
restarted. Setting this up using the UID of the speaker should still
make things more robust.

A thing that is still not taken into consideration is what should happen
when the speaker is not the coordinator of its group (this is also never
actually checked)