Refactor the sonos package
The tracks aren't used anymore, as they are slower than just having the
sonos speaker adding the playlist, which also makes then useless to
store. Remove the tracks from the card storage, and instead just track
the album and playlist IDs. This is also more accurate in general, as
these tracks will not be used when playing anyway (but rather what is on
the album/playlist according to deezer at that time)