Clean up albums/playlists
Make the code a little bit more generic (still some way to go here) and
clean up the CLI parameters. Store the title of the album/playlist in
the database, so that it is possible to see without re-fetching it from
the deezer API, and so that it is possible to have a human readable
version if deezer changes the ID for it (which happens from time to
time).